RNA: add a property flag to display exact integer values without fraction.
Add a flag to display exact integer values of a floating point field
without a fraction, for use in case when the value can technically be
fractional, but most commonly is supposed to be integer.

The fraction is discarded in the normal display mode and when copying
the value to clipboard, but not when editing to remind the user that
the field allows fractions.

Also, fix a precision issue when stepping down from 1 to 0 via the
left decrement button and step 100 results in a small nonzero value.

Differential Revision: https://developer.blender.org/D13753